acme-backend/pyproject.toml
[project]
name = "acme-backend"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = ["dagger-io"]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
